fixed the following issues: In the path_assemble function, the last line should be outside the for loop to correctly populate the path_param array.
In the path_position function, the line i = max(np.where(param > path.param)[0]) will raise an IndexError if param is less than the minimum value of path.param. It is recommended to add a check to ensure that param is within the range of path.param.

In the character class, the default values for the position, velocity, linear, and angular parameters are not valid numpy arrays. They should be written as np.array([0, 0]) instead of ([0, 0]).

Renamed character class to Character